Webpage/requirements.txt
azure-iot-hub==2.6.1
Flask==3.0.3
